+++ package.json
@@ -23,6 +23,9 @@
     "webpack-dev-server": "^3.11.2"
   },
   "dependencies": {
-    "phaser": "^3.52.0"
+    "bootstrap": "^4.6.0",
+    "jquery": "^3.5.1",
+    "phaser": "^3.52.0",
+    "popper.js": "^1.16.1"
   }
 }
