+++ src/Config/config.js
@@ -4,12 +4,13 @@ export default {
   type: Phaser.AUTO,
   width: 1000,
   height: 600,
+  parent: 'game-container',
   pixelArt: true,
   transparent: true,
   physics: {
     default: 'arcade',
     arcade: {
-      debug: true
+      debug: false
     }
   }
 };
