+++ src/style.css
@@ -1,0 +1,28 @@
+.bg-color-primary {
+    background-color: #e8e9fc;
+}
+
+.button-color {
+    background-color: #4d90cc;
+    cursor: pointer;
+}
+
+.button-color:hover {
+    text-decoration: none;
+}
+
+canvas {
+    width: 100%;
+    height: 100%;
+    object-fit: contain;
+    border: 1px solid #000;
+    border-radius: 8px;
+}
+
+input {
+    width: 50%;
+    box-shadow: 3px 10px 23px 0px rgb(0 0 0 / 75%);
+    border-radius: 5px;
+    height: 53px;
+    font-size: 2rem;
+}
